modules/location-consumer/requirements.txt
kafka-python==2.0.2
SQLAlchemy==1.3.19
GeoAlchemy2==0.8.4
Shapely==1.7.0